refactor&fix: crawling domain refactoring & fix crawling foriegn table
