+++ agaein_server/graphql/resolvers/crawling/queries.ts
@@ -1,36 +1,21 @@
-import { ApolloError } from 'apollo-server-errors';
-import { knex } from '../../database';
-import { readAccessToken } from '../../../common/auth/jwtToken';
-import { resourceLimits } from 'worker_threads';
+import { ApolloError } from 'apollo-server-express';
+import { getUserId } from '../../../common/auth/jwtToken';
+import { QueryCrawlingResultsArgs } from '../../types';
+import { getCrawlingHistory, getCrawlingResults } from './services';
 
 const crawlingQueries = {
-    crawlingResults: async (_: any, args: any) => {
-        const history = await knex('crawling_history').where('id', args.id).first();
-
-        if (history === undefined) {
-            throw new ApolloError(`CrawlingResults id(${args.id}) is not Exists`, 'BAD_USER_INPUT');
+    crawlingResults: async (_: any, crawlingRequest: QueryCrawlingResultsArgs) =>
+        getCrawlingResults(crawlingRequest.id),
+    crawlingHistory: async (_: any, __: any, context: any) => {
+        let userId: number;
+        const authorization: string = context.req.headers.authorization;
+        if (authorization && authorization.split(' ')[1]) {
+            userId = getUserId(authorization);
+        } else {
+            throw new ApolloError('No Authorization', 'BAD_USER_INPUT');
         }
 
-        return history.crawlingResults.results.map((result: any) => result.value).splice(0, 100);
-    },
-    crawlingHistory: async (_: any, args: any, context: any) => {
-        let userId = 1;
-        if (context.req.headers.authorization && context.req.headers.authorization.split(' ')[1]) {
-            const jwtToken = readAccessToken(context.req.headers.authorization.split(' ')[1]);
-            userId = (<any>jwtToken).userId;
-        }
-
-        const user = await knex('user').where('id', userId).first();
-
-        const rawHistories = await knex('crawling_history').where('user_id', userId);
-
-        const histories = rawHistories.map((history: any) => {
-            history.user = user;
-            history.crawlingResults = history.crawlingResults.results.map((result: any) => result.value);
-            return history;
-        });
-
-        return histories;
+        return await getCrawlingHistory(userId);
     },
 };
