+++ agaein_server/graphql/typedefs/index.graphql
@@ -50,5 +50,5 @@ type Mutation {
     updateReport(id: ID!, files: [Upload]!, report: UpdateReportInput!): Report!
     deleteReport(id: ID!, password: String): ID!
     # Crawling
-    crawling(baseInfo: CrawlingInput!, type: FINDING_TYPE): ID!
+    crawling(baseInfo: CrawlingInput!, type: FINDING_TYPE!): ID!
 }
